// src/models/IFlatDefect.ts
export interface IFlatDefect {
    imNumber?: number;
    imDescription?: string;
    imDescriptionWelsh?: string;
    itemNumber?: number;
    itemDescription?: string;
    itemDescriptionWelsh?: string;
    ref?: string;
    deficiencyText?: string;
    deficiencyTextWelsh?: string;
    forVehicleType?: string[];
}